// src/models/apflora/Adresse.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for apflora.adresse */
export type AdresseId = string & { __brand: 'apflora.adresse' };

/**
 * Represents the table apflora.adresse
 * Adressen, die in anderen Tabellen zugeordent werden können. Nicht zu verwechseln mit Konten, welche den Zugriff auf apflora.ch ermöglichen (Tabelle apflora.user)
 */
export default interface Adresse {
  /** Primärschlüssel */
  id: AdresseId;

  /** Vor- und Nachname */
  name: string | null;

  /** Strasse, PLZ und Ort */
  adresse: string | null;

  /** Telefonnummer */
  telefon: string | null;

  /** Email */
  email: string | null;

  /** Von wem wurde der Datensatz zuletzt geändert? */
  changed_by: string | null;

  /** Ist die Person freiwillige(r) Kontrolleur(in) */
  freiw_erfko: boolean | null;

  created_at: Date;

  updated_at: Date;
}

/**
 * Represents the initializer for the table apflora.adresse
 * Adressen, die in anderen Tabellen zugeordent werden können. Nicht zu verwechseln mit Konten, welche den Zugriff auf apflora.ch ermöglichen (Tabelle apflora.user)
 */
export interface AdresseInitializer {
  /**
   * Primärschlüssel
   * Default value: uuid_generate_v1mc()
   */
  id?: AdresseId;

  /** Vor- und Nachname */
  name?: string | null;

  /** Strasse, PLZ und Ort */
  adresse?: string | null;

  /** Telefonnummer */
  telefon?: string | null;

  /** Email */
  email?: string | null;

  /**
   * Von wem wurde der Datensatz zuletzt geändert?
   * Default value: NULL::character varying
   */
  changed_by?: string | null;

  /**
   * Ist die Person freiwillige(r) Kontrolleur(in)
   * Default value: false
   */
  freiw_erfko?: boolean | null;

  /** Default value: now() */
  created_at?: Date;

  /** Default value: now() */
  updated_at?: Date;
}

/**
 * Represents the mutator for the table apflora.adresse
 * Adressen, die in anderen Tabellen zugeordent werden können. Nicht zu verwechseln mit Konten, welche den Zugriff auf apflora.ch ermöglichen (Tabelle apflora.user)
 */
export interface AdresseMutator {
  /** Primärschlüssel */
  id?: AdresseId;

  /** Vor- und Nachname */
  name?: string | null;

  /** Strasse, PLZ und Ort */
  adresse?: string | null;

  /** Telefonnummer */
  telefon?: string | null;

  /** Email */
  email?: string | null;

  /** Von wem wurde der Datensatz zuletzt geändert? */
  changed_by?: string | null;

  /** Ist die Person freiwillige(r) Kontrolleur(in) */
  freiw_erfko?: boolean | null;

  created_at?: Date;

  updated_at?: Date;
}